# Speedup binding circuit params