<!--
⚠️ If you do not respect this template, your pull request will be closed.
⚠️ Your pull request title should be short detailed and understandable for all.
⚠️ Also, please add a release note file using reno if the change needs to be
  documented in the release notes.
⚠️ If your pull request fixes an open issue, please link to the issue.

- [ ] I have added the tests to cover my changes.
- [ ] I have updated the documentation accordingly.
- [ ] I have read the CONTRIBUTING document.
-->

### Summary

Speedup binding and returning circuit parameters by

* only sorting parameters when it's really needed 
* caching the sorted parameters

This should tackle the ~70x slowdown as presented in https://qiskit.github.io/qiskit/#circuit_construction.ParameterizedCircuitBindBench.time_bind_params?machine=dedicated-benchmarking-softlayer-baremetal&os=Linux%204.15.0-46-generic&ram=16GB&p-width=20&p-gates=2048&p-number%20of%20params=2048&commits=296a745b.

### Details and comments

On a 1000 parameter circuit binding 10 times currently takes
```
dict took 74.88096404075623s # bind via dict
array took 74.05472993850708s  # bind via array
```
with this PR its 
```
dict took 1.5851850509643555s 
array took 0.6309740543365479s
``` 

Fixes #5965