@Cryoris 2021-03-23
It would be nice to have a more uniform interface for accessing the parameters that would also allow `__len__` and `__contains__`, I agree! @molar-volume also suggested that while working on #5961.

It's not that straightforward though, since 
1. the parameter table doesn't know all parameters (global phase might be parameterized) 
2. the parameter view object was designed to be temporary and is currently only constructed upon returning the parameters
If we instead keep a `ParameterView` object internally (that also might do more efficient caching like you suggested in the comments), we could improve all these points. But this is likely better done in a follow-up, too 🙂