@kdk 2021-03-23
> It would be nice to have a more uniform interface for accessing the parameters that would also allow `__len__` and `__contains__`, I agree! @molar-volume also suggested that while working on #5961.

@Cryoris is there an issue to track these discussions so we can follow up once these PRs are merged? Can you open one if not?