@molar-volume 2021-03-24
I opened a draft suggesting changes to `ParameterView` and exploring options for simplifying `ParameterTable` as well https://github.com/Qiskit/qiskit-terra/pull/6074